Show version of plugin on top app bar
We now have a version file, with a simple semver string indicating the
plugin version.

We added a route in the Flask backend to read from this file and return
the contents.

We added a fetch call in the React frontend to GET this new route and
display the response in the UI.